Patch from Andy Ross:
Try the following patch to c172.xml, which makes the takeoff-rpm
modification I talked about*.  Hopefully things will be more to your
liking; the aircraft maxes out at 2100 RPM while stopped, then the
speed increases to something more like cruise as the airspeed grows.

If you look carefully, you'll actually see the RPM drop very slightly
before it starts increasing.  The physical reason for this is that the
blades are "unstalling".  As the flow attaches to them, they
experience a sharp increase in induced drag.  I was pretty pleased to
notice this little tidbit; it kinda validates the model in an obtuse
way.  That being said, I have *no* idea if this effect is noticeable
in a real aircraft.  Alex?