Introduced a new alias arrangement, using the 'include' attribute on the root PropertyNode element.  For example,
  --aircraft=c172

is now an alias for

  --aircraft=c172r

which is an alias for

  --aircraft=c172r-jsbsim

All JSBSim *-set.xml files have been renamed to include 'jsbsim'
explicitly in their names; the ones without 'jsbsim' are now aliases
to the default, whatever that may be.

Here are some new, simpler aircraft identifiers:

  --aircraft=cub
  --aircraft=747
  --aircraft=sopwithCamel
  --aircraft=c310-3d

and so on.

This system allows users to create new *-set.xml files by overriding
parts of existing ones rather than by cut-and-paste.