Lee Elliott:
I've taken the liberty of attaching a .tar.gz file containing a .3ds model of
a BAC-TSR2, a yasim config file based on the correct figures (where I could
find them) and the -set.xml and model.xml files to fly it.

I'm primarily a 3d'er and originally did the TSR2 for a picture I'm working
on but when I got fgfs running (Debian Linux) I couldn't resist loading it in
and trying to get it to fly.  The model was created in Realsoft3D and
exported as .3ds.

I've been able to tag the various sub-objects, to animate them but the export
process appears to 'flatten' any object hierarchy I set up and I'm guessing
this is necessary for sequential animations - I couldn't set up the correct
sequential rotations to properly bring the main u/c in.  Also, in real life,
there are several other u/c doors that should open and close in sequence to
get the gear in and apparently the sequence was quite complex.

On the ground though, it is as shown (so I didn't need to model the extra
doors for my picture anyway;)

It could do with some airbrakes too, both for the model and for the fdm.  As
with the extra u/c doors, I didn't need them for the picture and they haven't
been modelled.

As well as not being able to preserve object hierarchies when I export from
Realsoft3D's native object format to .3ds, I'm not able to preserve textures
or colour-mapping either, so the aircraft appears all white.

Hopefully, someone might like to add the extra doors and airbrakes, which
shouldn't be too difficult, and put some texturing on it - mostly white
anyway, for the prototypes, or a contemporary RAF scheme if someone wants to
pretend that it entered service.

The yasim fdm model, as said, cannot be regarded as accurate.  However, while
it is based on the specs for the real aircraft, where I could find them, the
measurments are probably only accurate to about 1 metre.  That's assuming I
was measuring the right things in the first place;)  Other bits that I wasn't
sure about i.e. flaps, ailerons etc. have been hacked out of the a4 or the
747.

It could do with some 'refinment' by people who know what they're doing, but
it seems to fly about right, or rather, as I'd imagine:)  (me want a
forward-looking ski-toe terrain avoidance radar:)

Anyway, I'm happy for the whole lot to be released under the same licence and
conditions as the rest of the fgfs stuff, either as a part of fgfs or by
anyone else who will also follow those same licence and condition terms.

I've also got a reasonable yasim b52 flying but no moveable bits on the model
yet, a vulcan with a similar simple 3d model using a grossly hacked c310
jbsim fdm (right numbers where I could find them but powered by a couple of
XLRs) and a Saunders-Roe SR45 Princess flying boat model and yasim fdm (can't
get it into the air with propellors but substituting equivilent jets (2.5x
factor) got it flying.  I've started on a EE/BAC Lighting FMk6 model but I'll
probably be doing a Fairchild A-10 and an Antonov An-225 first.

I figure this is the best way I can contribute to the fgfs project, and l'd
like to be able to offer something.