Lee Elliott
Here's a B52-H that I've done and it can be distributed under the same
conditions etc. as fgfs.  It seems to fly ok but I'm sure it could do with a
bit of adjustment - perhaps Andy R. will have a look and give it a few tweaks.

It wouldn't need too much work to convert it to an -A model (different
engines, taller fin etc.) and then it could be teamed up with the X15 :)

I still can't get textures on the models but it occurred to me that if
someone could apply blank texture maps to the models I could then fill in the
blanks, as it were.

I also noticed that there don't appear to be models included for the F-104,
F-15 & F-16 and I think I could probably do some and animate them for fgfs if
they'd be useful.

Re the A10 - Andy said he'd had a look at it and made a few changes to fix a
few things that I'd done incorrectly and asked if I wanted to keep tinkering
with it or should he roll his changes into the fgfs cvs?  I said to him that
I'd be happy to go with his changes as they'll be more realistic/accurate than
the stuff I've done.

I've certainly no problem with anyone doing anything with any of the stuff
I've done, so long as it's an improvement :)