Lee Elliott:
Here're a couple of updates.  Well, one's an update - the tsr2 and this now
has reasonable wing and hstab values.  The roll rate was very high and I
figured that as it was going to be a hack anyway, due to the all moving
tailplanes, I just made them very ineffective and linked them top the
tailplanes in the model file.  I think I may have over factored them a bit
too much but it's interesting watch them in chase view.  If the autopilot
roll-out-deg is dropped to about two it'll start oscillating, and from behind
it looks like it's paddling:)