Lee Elliott:
Here are some more updates, and bits and pieces.

The tsr2 is mostly a re-write, with more reasonable figures for the wing and
hstab.  I've also tried to reduce the poly count on the model whilst adding
the airbrakes and canopies.

Finally, I've set up some panels, based on the c310 vfr and mini panels.  The
vfr panels have a lot of additional instruments on them but impose quite a
high frame-rate hit on my system.  The mini panels also have some additional
instruments but seem to work without noticable penalty here.  I've set these
panels as the defaults for these aircraft but I'm not sure it's a good idea
to have the vfr panels by default due to the performance hit.  Using them as
default will should give some feedback though;)

To go with the new panels are a number of rough and ready instruments I
hacked, mostly out of one of the rpm gauges, but I've adapted the throttle
quadrent (jet-throttle-quadrant) to funstion for throttle, flaps and reheat.
All the gauge faces were based on an existing ffgfs instrument texture, so
there's no problem with copyright there.  All the other bits that I've done
may be distributed under the same conditions as fgfs.