Lee Elliott:
There's now another a10 variant - a ferry load configuration - no weapons but
3 x 600 us gal external tanks.  This seems to be the heaviest confguration
for an a10.  I also found a performance graph for the a10 in clean
configuration and it showed max speeds of 275 knots at 45000 ft and 398 knots
at 1200 ft.  As things seem to get a bit squiffy above 35000 ft, in my
experience, I used the 1200 ft figure.  This is pretty close the the figure
that Andy Ross found (380 knots at 0 ft) and I've also tried to incorporate
his mods into the new configs, with a better balanced wing/hstab settings.

After getting a good clean config, I added the extra masses/fuel
tanks/fueselages but didn't change any of the rest of the config, and it
still seems to work properly.  The external tanks were done as fuel tank
entries, with corresponding fuselage entries, whereas I'm not sure how to
handle the 'weight' entries in yasim yet so the weapon load is still done
with an increased total mass and ballast entries.

There'll be some stuff in the archives that haven't changed i.e. the cl & wl
a10 and seahawk models, but until I get using cvs for my own bits and pieces
it was easier to do the whole directory.  I hope this isn't a problem.

Finally, I've set up some panels, based on the c310 vfr and mini panels.  The
vfr panels have a lot of additional instruments on them but impose quite a
high frame-rate hit on my system.  The mini panels also have some additional
instruments but seem to work without noticable penalty here.  I've set these
panels as the defaults for these aircraft but I'm not sure it's a good idea
to have the vfr panels by default due to the performance hit.  Using them as
default will should give some feedback though;)

To go with the new panels are a number of rough and ready instruments I
hacked, mostly out of one of the rpm gauges, but I've adapted the throttle
quadrent (jet-throttle-quadrant) to funstion for throttle, flaps and reheat.
All the gauge faces were based on an existing ffgfs instrument texture, so
there's no problem with copyright there.  All the other bits that I've done
may be distributed under the same conditions as fgfs.