Patch from Julian Foad:
OK -- thanks.  I have now fixed up all the other XML files in the base
package that use "max" and "wrap".  Those affected are all in
Aircraft/Instruments/.  In these files I have also corrected the
indentation in some places, and corrected the frequency controls in
nav3.xml (which does not seem to be in use), and corrected a few obvious
hotspot position bugs.  Where the radios used wrapping to toggle the
ident button between 0 and 1, I have changed this to use
"property-toggle" instead.