Hello Curt,
here's an update to the b52 - it's just some relatively minor changes to the
fdm and panels.

I've been trying to improve the take-off and landing characteristics, or at
least get them to fit the pictures and film I've seen.

The best way to take off is by using the autopilot - at sea level just extend
the flaps, punch in the altitude hold on the autopilot (and heading if you're
lazy), and apply full power.  It'll also take of with the default the default
elevator trim of -0.07 but it 'staircases' a little bit - the auto pilot
smooths that out.

I've included some notes and observations about taking-off and landing in the
readme.

While I wouldn't claim any sort of accuracy for it, I hope it manages to get
some of the characteristics right.