Lee Elliott:
I've attached an update for the yf23 and I think I've included all the right
files - the model is unchanged (I think I'll eventually paint it in an
imaginary NASA scheme, perhaps a bit like the HIDEC F-15, until I can get
some newer pictures and see what they've done with it) so I've not included
it again.

The fdm still needs quite a bit of work - it can be a bit twitchy and wobbly
at times, and the auto-pilot roll-out and smooth need tightening up, but it's
flyable.  Haven't tried landing it yet though;)