Lee Elliott:
Here's a quick update to the yf23 fdm and auto-pilot settings.  It flies and
takes-off reasonably well - still haven't tried landing it yet;)  It can
still be a bit wobbly at times though and It's not very good at terrain
following either.

Taking off on auto-pilot is quite fun.  Set full flaps and CTRL-a (assuming
you're near sea level), hit F6 to make it go straight and apply power - about
50% until it rotates then apply full power.  Retract the gear as soon as it
gets off the ground and by the time you#ve done that, retract the flaps.
Quite entertaining from a tower view:)

It's not really an ideal a/c to model but apart from the aesthetics, which
appeal to me, I was curious about not having a rudder.

There's a new texture sheet too but it only makes the a/c a paler shade of
grey.