Lee Elliott:
I've been messing with the yf23 some more.  I think it flies a bit better now
and has more reasonable auto-pilot take-off characteristics - just apply full
power and go, and it doesn't go into such a steep climb.

Out of curiosity, I've also taken the rudder control input and linked it to
the stabilators/elevators and ailerons.  Now, with some slick joystick
manipulation it's possible to fly it at about 30 deg off-axis in more or less
level flight (actually, with full 'rudder' input, it looks nearer to 40 deg
to me and you end up tracing a very long curve).

If no rudder input is used it flies 'normally'.

I'd like to be able to 'tune' it so that it's more stable while transitioning
but I'm not sure if that's going to be possible.  It's also high-lighted some
problems I'm having with my Sidewinder joystick - when I use the rudder
control, the engine No.2 throttle jumps to 50% unless auto-throttle is
engaged (this also happens with the b52 too), and there seems to be some
interaction between the different axis if all three (roll, pitch and rudder)
are changed at the same time - difficult to be sure on that one though.

I'll probably un-map the throttle control from the joystick as I tend to use
the keyboard for engine control anyway, especially for landing.