Lee Elliott:
Here's an update to the yf23.  I've added a couple of small details to the
model and played with the textures a bit:)  Still no response from the chap
at NASA about sticking their logos on it yet.

I've also been fiddling with the fdm.  It had been bugging me that I couldn't
find any sign of air/speed brakes but just recently I found that braking is
done by moving the ailerons up and the flaps down, in opposition to each
other, which I thought was pretty neat really.

This was actually pretty easy to mimic - although there are no spoiler
surfaces defined in the YASim wing entry I was still able to map the spoiler
control input to the ailerons and flaps, and once I'd got the directions
right it seems to work - hit the spoilers and the ailerons move up, the flaps
drop and the energy worm shows that I'm losing speed.

The downside is that the flaps and the ailerons have to have similar lift and
drag characteristics, to balance each each other, and this means that it ends
up with the flaps not giving much lift and the brakes not having much affect
at low speeds.  If I could change the deflection rates then I could balance
them that way (and improve the rudder control too) - as it is, the deflection
for the flaps and ailerons is the same, just in opposite directions.  This
sounds like a job for interpolation tables but I haven't tried them yet and I
don't know if I can use them in this way in YASim.

The opposing flap/aileron techique is also used for manuevering e.g. by using
it on just one side but I haven't tried to incorporate that yet;)

Landing it is proving to be tricky - it gets down smoothly to about 20ft then
what I presume is ground effect kicks in and upsets the trim.  The only
answer appears to be to force the nose down at this point.  What's more, it
needs to be held down untill you trim the elevator out otherwise it'll
happily roll along on it's main gear, with it's nose in the air:)  There's
definitely not enough drag either.