I've tried to organise everything that the yf23 uses into it's Aircraft folder.  This has meant duplicating most of the instruments it uses but it should be easier to maintain and once I've worked through all the other a/c you'll be able to remove a lot of the cruft I'm responsible for in the generic instrument folders.
The only files that now don't reside in the yf23 folder are the 'set' file and
the yasim config.

The stuff in this archive should replace the entire existing yf23 folder and
the two 'set' and yasim files (unless I've screwed up somewhere;)

There are some small changes to the fdm.

I also had this rather silly idea, so I had to give it a go;)  Have a look at
the ground below the a/c.

It doesn't track the Sun, of course, so it stays directly below the a/c.  It
just uses a couple of rotations to keep it flat, and a translation to keep it
on the ground.  It seemed strange curious that I had to use a translation
factor of 0.3 instead 1.0 to keep it in the right place.