Lee Elliott:
The base a10 yasim config wasn't working because of the control structure
changes so I've fixed that.  The other yasim configs are also included
because I forgot to omit them from the archive but they're not updated or
changed.

All the a10 stuff needs badly re-doing though and it's a bit of a mess - the
yasim configs are all similar but I don't know which ones had Andy Ross's
mods.  I've also got to either forget the weapon load version or get it
working with weight entries so that the rest of the configs can be
indentical.