Alex Perry: Since several people on the list have been asking questions about this Saitek joystick and the comments at the top of this xml file in cvs were wrong and confusing, I am posting this update with the following changes: 1.  Removed the confusing/wrong comments. 2.  Added throttle bindings for engines 2 thru 7 (required for the b52). 3.  Changed the hat axis to rotate the view continuously (required to see the mag switch in the p51d and just a lot nicer with any 3d cockpit).
This last change makes the hat point the line of sight more like moving your head.  This is the way the CH pro-yoke-USB is set up.

If you are not used to this set up and you use aircraft with 2d pannels, you will likely need to use shift-nuberpad-8 to get the original forward view once you look arround.