Lee Elliot:
I've added reheat to the engines to simulate water injection and then tweaked
it so that it's just about possible to take off with the full fuel load/ferry
configuration (a little under the mtow - 420,000lbs vs 450,000lbs).

It's interesting to note that the normal combat tow for the F model was
291,570lbs - equivilent to a fuel load of about 0.46 (this would actually
leave you with more fuel than in real life because some of that weight would
be the weapons load but here it's fuel)

Anyway, I've added all this info to the readme and explained some of the
reasoning behind the fdm there.

It still doesn't reach the correct altitudes, and certainly not with a full
fuel load, but it starts getting into the right ball park at combat tow.