Lee Elliott:
I've added reheat to the engines to simulate water injection and then tweaked
it so that it's just about possible to take off with the full fuel load/ferry
configuration (a little under the mtow - 420,000lbs vs 450,000lbs).

It's interesting to note that the normal combat tow for the F model was
291,570lbs - equivilent to a fuel load of about 0.46 (this would actually
leave you with more fuel than in real life because some of that weight would
be the weapons load but here it's fuel)

Anyway, I've added all this info to the readme and explained some of the
reasoning behind the fdm there.

It still doesn't reach the correct altitudes, and certainly not with a full
fuel load, but it starts getting into the right ball park at combat tow.

I've also replaced the transparent instruments on the mini panel with some
text/digital ones.  I wanted more accurate numbers when I was monitoring the
effects of fdm changes and the dials just didn't give the detail.  They
aren't perfect and be difficult to see against certain backgrounds and in
certain lighting conditions but what the hey - nearly everything I do is a
bit of an experiment;)

Something I've noticed is that since I've started identifying flight surfaces
and gear components individually I've noticed some problems with the flaps
and gear in replay.  I address the left and right flaps independently so they
have left&right-flap-pos-norms and don't get controlled by the replay
mechanism which must refer to the single flap-pos-norm.  There's a similar
thing with the gear - I have to identify them individually so I can animate
them properly so on the b52 there're six gear elements but only the first
three are actioned in replay.

Personally, I think that the only way around this will be to provide
control-overrides for all these items, much like there're the AP control
overrides for throttle etc.