David Culp:
The radar instrument uses the above three items, and applies a scale factor to
the x-shift and y-shift in order to match the instrument's scale.  Changing
the display scale can be done entirely in the XML code for the instrument.
Right now it's set up only to display a 40 mile scale.

The radar is an AWACS view, which is not very realistic, but it is useful and
demonstrates the technology.  With just a little more work I can get a HUD
marker.  All I need to do there is make a bank angle adjustment to the
current values.

This works great with one target.  For two or more targets the radar
instrument will have to know the numbering of the aircraft model properties.
This isn't implemented yet.

David Culp:
I couldn't stop.  Here's a better radar instrument.  It has:

1)  range select knob 20 and 40 nm (not clickable.. yet)
2)  target altitude readout at lower right
3)  target disappears when range exceeds 43 nm
4)  range ring values now are read from instrumentation/radar/range
5)  instrumentation/radar/range is preset in the *-set.xml file to 40 nm

The next step would be a clickable range selection.  The problem here is that
the instrument currently displays the "blip" only if the target's range is
less then 43 nm.  If the range scale is decreased to 20 nm, then the "blip"
will show past the edges of the instrument.  I might need to make another
instrument for the 20 nm scale to make that work.