David Culp:
OV10-jsbsim-set.xml  has many changes.  It creates a thunderstorm over
woodside VOR.  It removes the lowest cloud layer so that the storm can be
seen.  It sets the surface wind speed to zero, which may help stop the
"drifting left" complaints. It sets the radar range to 40 miles.


The weather radar instrument is corrected.