Eric L. Hathaway:
The following patch updates the ThrustMaster FCS joystick configuration.  I have "Nasal-ized" the joystick bindings, drawing ideas from the Cyborg-Gold-3d-USB configuration file.  I also changed some of the bindings, so the joystick setup is more like the default four-axis-joystick config.  When I submitted the original config file, I had the hat switch bound to the rudder and elevator trim.  Since the vast majority (all?) of the other joystick configs use the hat switch to control view direction, I think it would be best for the defaults for this joystick to conform to the rest in order to obey the "principle of least surprise" for the unsuspecting user.