Mathias Fröhlich:
Here is an other one to the data directory. It updates the carrier launchbar
command and catapult launch command to reset itself if the key is no longer
pressed. That fixes the odd behavour, that once a carrier launch happened, a
second lauch happens no longer immediately at arresing time.