Stuart Buchanan:
joystick configuration for a Wingman Force 3D USB.

mfranz:
This driver is supposed to work for the non-USB versions, too, so
I'm replacing the non-USB driver with this one, and add its <name>s
here. If it turns out that we need a separate driver, it's better to
derive it from this file, as this is adapted to the newest methods
(nasal wrappers, etc.), while the old driver was only partly functional.