Changes from Dave Perry:
I put 3 patches and a sound file in this tar.  The c172p... patch
removes an unnecessary patch I submitted to the c172p.  It has no affect
on performance.  The Instruments-3d... patch adds a working skid ball to
the tc and switches the vor xmls to using params so it is easier to
edit.  It also constrains the glide-slope needle to -32 to 32 degrees.
The pa24-250... patch fixes an edit error (corrects a misnamed switch
"click", no change in performance) and adds a "squish" sound to the
primer pump action.