Rob Oates:
New textures update:

Include a new Town texture (now just 1 texture...originally was 3) This
looks REALLY good and it goes will with all terrain types, and best of all
it's a single texture...town2, and town3 are nolonger apart of the download
package.

A new sand texture (2 textures...for testing) which I hope John? can test
this because he's very familar with desert terrain...I've only seen it in
pictures.

And an improved Drycrop textures (now 4 Textures...was 3) now the texture
looks more rural, you can best see the texture from KTPA.

If the feedback is good on the desert texture I will do a full clean set
so you can delete the previous texture folder and have a clean set of all
the Flightgear textures.