In mouse view mode <=> one could already move the view left/right/up/down when dragging the mouse with middle mouse button (MMB) pressed (reset with MMB+LMB):
Add support for forward/backward moving (MMB+KbdCtrl). With this one can,
for example, read what's written on the backside of the cup of coffee
on top of the screen in the E3B's operator room, or walk around between
the passenger seats in the fokker70 searching for the toilet.  :-)