first stab at file-select dialog. Initial target dir can be set in /sim/gui/dialogs/file-select/directory, the selected file's full path is returned in /sim/gui/dialogs/file-select/directory. There will be a gui.file_selector class that allows to have more than one of these dialogs open, where each of them reports the result to a callback function.
The dialog toggles dotfile display with ctrl-click on the "." entry,
and switches to FG_ROOT and FG_HOME with ctrl-click and shift-click on
the ".." entry.

This dialog is ATM not used anywhere in fgfs, but might be needed by
add-ons, much like the xml.nas code. Note that backslashes are converted
to slashes, which should ensure that no characters can be sneaked in
on UNIX systems. This breaks files containing escaped characters on
MS Windows. Fix for that on request. :-)  Special characters (like German
umlauts) aren't displayed correctly, but properly returned. (Depending
on the plib font.)