Torsten Dreyer:
This patch for the 3d artificial horizon enables the movement of the
reference bar like the 2d instrument and corrects the wrong display of
pitch.

The original version already has a reference bar object in the .ac file
called bezel, but it is not bound to a animation. And the pitch is
indicated approx. 25% low so a pitch of +10deg is indicated as approx.
7deg.