- remove property browser binding from the <SPACE> key - rename prop-key-handler.nas to prop_key_handler.nas  (the hyphens were   intentional, exactly to *make* using it as namespace less inviting,   but times change ...) - add property browser binding to the '/'-key for when the property key handler   is turned off (/sim/input/property-key-handler=0). If it's on, use /: or   /<property>: to open the browser. - run keyboard event listener only when property key handler is active
I hope that the '/' key can keep this binding even after a keyboard review.
The '/' is just the most natural key for dealing with properties, and it's
far less prominent on non-US-keyboards (e.g. Shift-7 on German keyboards),
so it's not really very well suited for important aircraft functions, anyway.
But I don't insist. :-)