start vi-like multiple-key commands with ':'
Note that this key does sometimes fail to work. This is caused by the
way our keyboard works ATM. On an (US)English keyboard : and ; are on
the same key. If you do  Press-Shift >> Press-: >> Release-Shift -> Release-:
then not the ':' is released, but the ';'. The ':' remains registered as
pressed, so the next press fails. Another press restores the function.
We fixed this for the letter keys, as it is safe to assume that 'a' and
'A' are on the same key. Unfortunately, we can't make such assumptions
for non-letter keys. We should really track the HW-key numbers, not
the symbols printed on them!