Issue 209: fix up inheritance of color mode uniform in material animation effect
The overriding of the color mode uniform in the default model effect
wasn't working. It's difficult to override the individual properties
of techniques, especially if their names aren't unique. So, I changed
the model-default effect to use a parameter for the color-mode
uniform; material-off.eff and material-diffuse.eff override that.