Location-in-Air: pick nearest VOR/NDB (fix#228)
When entering a duplicate VOR or NDB id into the Location-in-Air
dialog, the first entry in the database was picked. This patch
assumes that the station with the given ID nearest to the current
position was requested.
A better fix would probably be to present the user a list of found
stations - that's something for the future.