Refactor weather-utility.nas
A bug in weather-utility.nas caused constant computation
of wave parameters due to the fact that the wind speed
was ignored.

That code was ugly anyway, so it has been replaced by
property rules which don't create work for the nasal
garbage collector.

All required properties for the wave shader are now in
interpolator.xml, the snow-line computation lives in
metarinterpolator.xml