Add optional <page> and <marker> tags to checklists
Checklists may now be split into individual sections made up
of <pages> containing the <item> tags.  Each page is displayed
individually with Previous/Next buttons to allow navigation.

Checklists items also support <marker> tags.  These display
a marker when a ? button is pressed next to the checklist item.
The format is identical to that of the tutorial system.