Implemented a documentation compiler, initially for Aircraft
Forum thread: http://forum.flightgear.org/viewtopic.php?f=72&t=25119
The documentation compiler will iterate over all Aircraft and generate
LaTeX files for each aircraft. It will then compile them to a pdf. Please
refer to the forum thread for an example of such a pdf file.

It will also generate a separate pdf file just for the checklists.

Currently, the compiler uses the checklist files and additional files in a
Docs/ subdirectory of the aircraft.

To use this script, you need python > 2.7 and a recent version of pdflatex
with hyperref support installed.