Checklists: Fix bug where first checklist has multiple pages
The change introduced in commit 8c7fc119 to go to the last checklist
on open causes a problem when the first checklist has multiple pages.
If the last opened checklist did not have multiple pages, the previous
and next page buttons are still shown.

This commit runs the same binding as used by the combo box to reset
the visibility of these buttons immediately before the page is
displayed.